refactor: move ticket from domain to view
