+++ app/src/main/java/woowacourse/movie/view/ticket/TicketActivity.kt
@@ -9,8 +9,8 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import woowacourse.movie.R
-import woowacourse.movie.domain.ticket.Ticket
 import woowacourse.movie.view.reservation.ReservationActivity
+import woowacourse.movie.view.ticket.model.Ticket
 
 class TicketActivity : AppCompatActivity() {
     private lateinit var ticket: Ticket
